Refactor webhook helpers to satisfy Next.js route constraints
