+++ app/api/kiwify/webhook/route.ts
@@ -5,6 +5,13 @@ import { NextResponse } from 'next/server';
 import { createClient, type SupabaseClient } from '@supabase/supabase-js';
 import * as crypto from 'crypto';
 import { readEnvValue } from '../../../../lib/env';
+import { deepWalk, pickByKeys } from './utils';
+import {
+  extractTrafficSource,
+  extractTrackingParams,
+  hasUrlTrackingParams,
+  mergeCheckoutUrlWithTracking,
+} from './traffic';
 
 // ==== ENVS OBRIGATÓRIAS ====
 // ==== ENVS OPCIONAIS ====
@@ -21,91 +28,6 @@ const STRICT_SIGNATURE =
 
 // ---------- Helpers ----------
 const EMAIL_RE = /^[^\s@]+@[^\s@]+\.[^\s@]+$/i;
-
-function deepWalk(
-  obj: any,
-  visit: (k: string, v: any, path: string[]) => boolean,
-  path: string[] = []
-): boolean {
-  if (obj === null || obj === undefined) return false;
-  if (Array.isArray(obj)) {
-    for (let i = 0; i < obj.length; i++) {
-      if (deepWalk(obj[i], visit, [...path, String(i)])) return true;
-    }
-    return false;
-  }
-  if (typeof obj === 'object') {
-    for (const [k, v] of Object.entries(obj)) {
-      if (visit(k, v, [...path, k]) || deepWalk(v, visit, [...path, k])) return true;
-    }
-    return false;
-  }
-  return false;
-}
-
-function getValueByPathCaseInsensitive(obj: any, path: string[]): any {
-  let current: any = obj;
-
-  for (const rawSegment of path) {
-    if (current === null || current === undefined) return undefined;
-
-    const segment = rawSegment.toLowerCase();
-
-    if (Array.isArray(current)) {
-      const index = Number(segment);
-      if (Number.isInteger(index) && index >= 0 && index < current.length) {
-        current = current[index];
-        continue;
-      }
-      return undefined;
-    }
-
-    if (typeof current === 'object') {
-      const entries = Object.entries(current as Record<string, any>);
-      const match = entries.find(([key]) => key.toLowerCase() === segment);
-      if (!match) return undefined;
-      current = match[1];
-      continue;
-    }
-
-    return undefined;
-  }
-
-  return current;
-}
-
-function pickByKeys(
-  obj: any,
-  keys: string[],
-  test?: (v: any) => boolean
-): any | null {
-  const normalizedKeys = keys.map((key) => key.toLowerCase());
-  const pathKeys = keys
-    .map((key) => key.trim())
-    .filter((key) => key.includes('.'))
-    .map((key) => key.split('.').filter(Boolean));
-
-  for (const path of pathKeys) {
-    const value = getValueByPathCaseInsensitive(obj, path);
-    if (value !== undefined && (test ? test(value) : value != null)) {
-      return value;
-    }
-  }
-
-  let found: any = null;
-  deepWalk(obj, (k, v) => {
-    if (
-      normalizedKeys.some((key) => !key.includes('.') && k.toLowerCase() === key) &&
-      (test ? test(v) : v != null)
-    ) {
-      found = v;
-      return true;
-    }
-    return false;
-  });
-  return found;
-}
-
 function findEmailDeep(obj: any): string | null {
   const byKey = pickByKeys(
     obj,
@@ -323,764 +245,6 @@ function findCheckoutUrlDeep(obj: any): string | null {
   return best;
 }
 
-function parseCheckoutSearchParams(url: string | null | undefined): URLSearchParams | null {
-  if (!url) return null;
-  try {
-    return new URL(url).searchParams;
-  } catch {
-    try {
-      return new URL(url, 'https://example.com').searchParams;
-    } catch {
-      return null;
-    }
-  }
-}
-
-const UTM_PARAM_PATHS: Record<string, string[]> = {
-  utm_source: ['utm_source', 'utmSource', 'trackingparameters.utm_source'],
-  utm_medium: ['utm_medium', 'utmMedium', 'trackingparameters.utm_medium'],
-  utm_campaign: ['utm_campaign', 'utmCampaign', 'trackingparameters.utm_campaign'],
-  utm_content: ['utm_content', 'utmContent', 'trackingparameters.utm_content'],
-  utm_term: ['utm_term', 'utmTerm', 'trackingparameters.utm_term'],
-  utm_id: ['utm_id', 'utmId', 'trackingparameters.utm_id'],
-  utm_source_platform: [
-    'utm_source_platform',
-    'utmSourcePlatform',
-    'trackingparameters.utm_source_platform',
-  ],
-  utm_platform: ['utm_platform', 'utmPlatform', 'trackingparameters.utm_platform'],
-  utm_channel: ['utm_channel', 'utmChannel', 'trackingparameters.utm_channel'],
-};
-
-type TrackingParamMap = Record<string, string>;
-
-function cleanTrackingValue(value: any): string {
-  return typeof value === 'string' ? value.trim() : '';
-}
-
-function extractTrackingParams(body: any): TrackingParamMap {
-  const params: TrackingParamMap = {};
-
-  for (const [target, paths] of Object.entries(UTM_PARAM_PATHS)) {
-    const value = pickByKeys(body, paths, (v) => typeof v === 'string' && v.trim().length > 0);
-    const cleaned = cleanTrackingValue(value);
-    if (cleaned) {
-      params[target] = cleaned;
-    }
-  }
-
-  return params;
-}
-
-function hasTrackingParams(params: TrackingParamMap): boolean {
-  return Object.keys(params).length > 0;
-}
-
-function hasUrlTrackingParams(url: string | null | undefined): boolean {
-  if (!url) return false;
-  try {
-    const parsed = new URL(url);
-    let found = false;
-    parsed.searchParams.forEach((_, key) => {
-      if (!found && key.toLowerCase().startsWith('utm_')) {
-        found = true;
-      }
-    });
-    if (found) {
-      return true;
-    }
-  } catch {
-    return false;
-  }
-  return false;
-}
-
-function sameUrlBase(a: string | null | undefined, b: string | null | undefined): boolean {
-  if (!a || !b) return false;
-  try {
-    const urlA = new URL(a);
-    const urlB = new URL(b);
-    return urlA.origin === urlB.origin && urlA.pathname === urlB.pathname;
-  } catch {
-    return false;
-  }
-}
-
-function hasParamCaseInsensitive(params: URLSearchParams, key: string): boolean {
-  const lower = key.toLowerCase();
-  let match = false;
-  params.forEach((_, existingKey) => {
-    if (!match && existingKey.toLowerCase() === lower) {
-      match = true;
-    }
-  });
-  return match;
-}
-
-function mergeCheckoutUrlWithTracking(
-  baseUrl: string | null,
-  params: TrackingParamMap,
-  fallbackUrl: string | null,
-): string | null {
-  if (!baseUrl) {
-    return fallbackUrl;
-  }
-
-  if (!hasTrackingParams(params)) {
-    if (
-      fallbackUrl &&
-      fallbackUrl !== baseUrl &&
-      hasUrlTrackingParams(fallbackUrl) &&
-      !hasUrlTrackingParams(baseUrl) &&
-      sameUrlBase(fallbackUrl, baseUrl)
-    ) {
-      return fallbackUrl;
-    }
-    return baseUrl;
-  }
-
-  try {
-    const parsed = new URL(baseUrl);
-    let changed = false;
-
-    for (const [key, value] of Object.entries(params)) {
-      const trimmed = value.trim();
-      if (!trimmed) continue;
-      if (hasParamCaseInsensitive(parsed.searchParams, key)) continue;
-      parsed.searchParams.set(key, trimmed);
-      changed = true;
-    }
-
-    return changed ? parsed.toString() : baseUrl;
-  } catch {
-    return baseUrl;
-  }
-}
-
-const MANUAL_TRACKING_PARAM_KEYS = ['rb_manual'];
-
-function detectManualReminder(params: URLSearchParams | null): string | null {
-  if (!params) return null;
-
-  for (const key of params.keys()) {
-    const normalizedKey = key.trim().toLowerCase();
-    if (MANUAL_TRACKING_PARAM_KEYS.includes(normalizedKey)) {
-      const value = params.get(key);
-      if (typeof value === 'string') {
-        const trimmed = value.trim().toLowerCase();
-        if (trimmed) {
-          return trimmed;
-        }
-      }
-    }
-  }
-
-  return null;
-}
-
-function normalizeTrafficString(value: string): string {
-  return value
-    .trim()
-    .toLowerCase()
-    .normalize('NFD')
-    .replace(/[\u0300-\u036f]/g, '')
-    .replace(/[^a-z0-9]+/g, ' ')
-    .trim();
-}
-
-type TrafficContext = {
-  tokens: Set<string>;
-  combined: string[];
-  paramKeys: Set<string>;
-  paramTokens: Set<string>;
-};
-
-type ChannelHintGroup = {
-  channel: string;
-  hints: string[];
-  strictTokens?: string[];
-};
-
-const CHANNEL_HINT_GROUPS: ChannelHintGroup[] = [
-  {
-    channel: 'tiktok',
-    hints: [
-      'tiktok',
-      'ttclid',
-      'ttad',
-      'ttads',
-      'ttadgroup',
-      'ttcampaign',
-      'tiktokads',
-      'tt pixel',
-      'bt pixel',
-    ],
-    strictTokens: ['tiktok'],
-  },
-  {
-    channel: 'instagram',
-    hints: [
-      'instagram',
-      'instagramads',
-      'igads',
-      'igad',
-      'instaads',
-      'instaad',
-      'igstory',
-      'igreels',
-      'igcampaign',
-    ],
-    strictTokens: ['instagram', 'insta', 'ig'],
-  },
-  {
-    channel: 'facebook',
-    hints: [
-      'facebook',
-      'facebookads',
-      'fbads',
-      'fbadset',
-      'fbcampaign',
-      'fbclid',
-      'meta ads',
-      'metaads',
-      'metapixel',
-      'metacampaign',
-    ],
-    strictTokens: ['facebook', 'fb', 'meta'],
-  },
-  {
-    channel: 'google',
-    hints: ['google', 'googleads', 'adwords', 'gads', 'gclid', 'sem', 'searchads'],
-    strictTokens: ['google'],
-  },
-  {
-    channel: 'bing',
-    hints: ['bing', 'bingads', 'msclkid', 'microsoft'],
-    strictTokens: ['bing'],
-  },
-  { channel: 'taboola', hints: ['taboola'], strictTokens: ['taboola'] },
-  {
-    channel: 'kwai',
-    hints: ['kwai', 'kwaiadid', 'kwaiads'],
-    strictTokens: ['kwai'],
-  },
-  {
-    channel: 'pinterest',
-    hints: ['pinterest', 'pinads'],
-    strictTokens: ['pinterest'],
-  },
-  {
-    channel: 'snapchat',
-    hints: ['snapchat', 'snap ads', 'snapads', 'snap'],
-    strictTokens: ['snapchat', 'snap'],
-  },
-  {
-    channel: 'twitter',
-    hints: ['twitter', 'xads', 'x com', 'x.com', 'twitterads'],
-    strictTokens: ['twitter'],
-  },
-  { channel: 'linkedin', hints: ['linkedin'], strictTokens: ['linkedin'] },
-  {
-    channel: 'youtube',
-    hints: ['youtube', 'youtubeads', 'ytads', 'yt campaign'],
-    strictTokens: ['youtube', 'yt'],
-  },
-  {
-    channel: 'email',
-    hints: ['email', 'newsletter', 'mailing'],
-    strictTokens: ['email'],
-  },
-  {
-    channel: 'whatsapp',
-    hints: ['whatsapp', 'zap', 'wpp', 'whats'],
-    strictTokens: ['whatsapp', 'zap', 'wpp'],
-  },
-];
-
-function buildTrafficContext(values: string[], params: URLSearchParams | null): TrafficContext {
-  const tokens = new Set<string>();
-  const combined: string[] = [];
-  const paramKeys = new Set<string>();
-  const paramTokens = new Set<string>();
-
-  const addValue = (raw: string | null | undefined) => {
-    if (!raw) return;
-    const normalized = normalizeTrafficString(String(raw));
-    if (!normalized) return;
-    combined.push(normalized);
-    for (const token of normalized.split(' ')) {
-      if (token) tokens.add(token);
-    }
-  };
-
-  for (const value of values) {
-    addValue(value);
-  }
-
-  if (params) {
-    for (const [key, value] of Array.from(params.entries())) {
-      const lowerKey = key.toLowerCase();
-      paramKeys.add(lowerKey);
-      addValue(lowerKey);
-      addValue(value);
-      for (const token of lowerKey.split(/[^a-z0-9]+/)) {
-        if (token) paramTokens.add(token);
-      }
-      for (const token of value.toLowerCase().split(/[^a-z0-9]+/)) {
-        if (token) paramTokens.add(token);
-      }
-    }
-  }
-
-  return { tokens, combined, paramKeys, paramTokens };
-}
-
-function hasTrafficHint(ctx: TrafficContext, hints: string[]): boolean {
-  for (const hint of hints) {
-    const normalized = normalizeTrafficString(hint);
-    if (!normalized) continue;
-    const tokens = normalized.split(' ');
-    if (tokens.some((token) => ctx.tokens.has(token) || ctx.paramTokens.has(token))) {
-      return true;
-    }
-    const compact = normalized.replace(/ /g, '');
-    if (compact && ctx.paramKeys.has(compact)) {
-      return true;
-    }
-    if (ctx.paramKeys.has(hint.toLowerCase())) {
-      return true;
-    }
-    for (const value of ctx.combined) {
-      if (value.includes(normalized)) return true;
-    }
-  }
-  return false;
-}
-
-function contextHasToken(ctx: TrafficContext, token: string): boolean {
-  const normalized = normalizeTrafficString(token);
-  if (!normalized) return false;
-  if (ctx.tokens.has(normalized) || ctx.paramTokens.has(normalized)) {
-    return true;
-  }
-  const compact = normalized.replace(/ /g, '');
-  if (compact && ctx.paramKeys.has(compact)) {
-    return true;
-  }
-  if (ctx.paramKeys.has(token.toLowerCase())) {
-    return true;
-  }
-  return false;
-}
-
-function detectChannelFromContext(ctx: TrafficContext): string | null {
-  for (const group of CHANNEL_HINT_GROUPS) {
-    const strictMatch = group.strictTokens?.some((token) => contextHasToken(ctx, token));
-    if (strictMatch || hasTrafficHint(ctx, group.hints)) {
-      return group.channel;
-    }
-  }
-  return null;
-}
-
-function joinTrafficParts(parts: Array<string | null | undefined>): string | null {
-  const normalized = parts
-    .map((part) => (typeof part === 'string' ? part : null))
-    .filter((part): part is string => !!part && !!part.trim())
-    .map((part) =>
-      part
-        .trim()
-        .toLowerCase()
-        .normalize('NFD')
-        .replace(/[\u0300-\u036f]/g, '')
-        .replace(/[^a-z0-9]+/g, '.')
-        .replace(/\.+/g, '.')
-        .replace(/^\.+|\.+$/g, '')
-    )
-    .filter(Boolean);
-
-  if (!normalized.length) return null;
-  return normalized.join('.');
-}
-
-function resolveChannelFromValue(value: string): string | null {
-  const normalized = normalizeTrafficString(value);
-  if (!normalized) return null;
-
-  const context = buildTrafficContext([value], null);
-  const channelFromHints = detectChannelFromContext(context);
-  if (channelFromHints) return channelFromHints;
-  if (/\baffiliate\b/.test(` ${normalized} `) || /\bafiliad[oa]\b/.test(` ${normalized} `) || /\bparceria\b/.test(` ${normalized} `)) {
-    return 'affiliate';
-  }
-
-  const padded = ` ${normalized} `;
-  if (/\btiktok\b/.test(padded) || normalized.includes('tiktokad')) return 'tiktok';
-  if (/\binstagram\b/.test(padded)) return 'instagram';
-  if (/\bfacebook\b/.test(padded) || /\bmeta\b/.test(padded) || padded.includes(' fb ')) return 'facebook';
-  if (/\bgoogle\b/.test(padded) || /\badwords\b/.test(padded) || /\bgads\b/.test(padded)) return 'google';
-  if (/\bbing\b/.test(padded) || /\bmicrosoft\b/.test(padded)) return 'bing';
-  if (/\bkwai\b/.test(padded)) return 'kwai';
-  if (/\btaboola\b/.test(padded)) return 'taboola';
-  if (/\bpinterest\b/.test(padded)) return 'pinterest';
-  if (/\bsnap(chat)?\b/.test(padded) || padded.includes(' snap ')) return 'snapchat';
-  if (/\btwitter\b/.test(padded) || /\bxads?\b/.test(padded) || normalized.includes('x.com')) return 'twitter';
-  if (/\blinked?in\b/.test(padded)) return 'linkedin';
-  if (/\byoutube\b/.test(padded)) return 'youtube';
-  if (/\bwhatsapp\b/.test(padded) || padded.includes(' wa ')) return 'whatsapp';
-  if (/\bemail\b/.test(padded) || normalized.includes('newsletter')) return 'email';
-  if (/\baffiliate\b/.test(padded) || /\bafiliad[oa]\b/.test(padded) || /\bparceria\b/.test(padded)) return 'affiliate';
-  if (/\borganic\b/.test(padded)) return 'organic';
-  if (/\bdirect\b/.test(padded)) return 'direct';
-
-  const slug = normalized.replace(/[^a-z0-9]+/g, '.').replace(/\.+/g, '.').replace(/^\.+|\.+$/g, '');
-  return slug || null;
-}
-
-function detectTrafficClassification(ctx: TrafficContext, params: URLSearchParams | null): string {
-  const affiliateTokens = [
-    'affiliate',
-    'affiliates',
-    'afiliado',
-    'afiliada',
-    'afiliados',
-    'afiliadas',
-    'afiliacao',
-    'parceria',
-    'parceiro',
-    'parceira',
-  ];
-  if (hasTrafficHint(ctx, affiliateTokens)) return 'affiliate';
-
-  const emailTokens = ['email', 'newsletter', 'mailing'];
-  if (hasTrafficHint(ctx, emailTokens)) return 'email';
-
-  const referralTokens = ['referral', 'indicacao', 'indication', 'referencia', 'referido'];
-  if (hasTrafficHint(ctx, referralTokens)) return 'referral';
-
-  const organicTokens = ['organic', 'organico', 'gratis', 'free', 'seo', 'blog', 'conteudo', 'content', 'direct'];
-  const hasOrganicHint = hasTrafficHint(ctx, organicTokens);
-
-  const paidTokens = [
-    'paid',
-    'pago',
-    'midia paga',
-    'midiapaga',
-    'trafego pago',
-    'trafegopago',
-    'ads',
-    'ad',
-    'cpc',
-    'cpa',
-    'cpp',
-    'cpm',
-    'ppc',
-    'display',
-    'remarketing',
-    'retargeting',
-    'sponsored',
-    'paid social',
-    'paid search',
-  ];
-
-  const paidParamIndicators = ['fbclid', 'gclid', 'msclkid', 'kwai', 'adset', 'adgroup', 'campaign_id', 'ad_id'];
-  const hasPaidParam = params
-    ? Array.from(params.keys()).some((key) =>
-        paidParamIndicators.some((indicator) => key.toLowerCase().includes(indicator))
-      )
-    : false;
-  const hasPaidHint = hasTrafficHint(ctx, paidTokens);
-
-  const tiktokTokens = ['tiktok', 'ttclid', 'ttad', 'ttadgroup', 'tiktokads'];
-  const hasTiktokHint = hasTrafficHint(ctx, tiktokTokens);
-
-  if (hasOrganicHint) return 'organic';
-
-  if (hasTiktokHint && !hasPaidHint && !hasPaidParam) {
-    return 'organic';
-  }
-
-  if (hasPaidHint || hasPaidParam) return 'paid';
-
-  if (hasTiktokHint) return 'organic';
-
-  return 'unknown';
-}
-
-export function extractTrafficSource(
-  body: any,
-  checkoutUrl: string | null,
-  existingTrafficSource: string | null
-): string | null {
-  const params = parseCheckoutSearchParams(checkoutUrl);
-  const manualReminder = detectManualReminder(params);
-
-  const sources: string[] = [];
-  const mediums: string[] = [];
-  const hints: string[] = [];
-
-  const addValue = (collection: string[], value: any) => {
-    if (typeof value !== 'string') return false;
-    const trimmed = value.trim();
-    if (!trimmed) return false;
-    if (collection.includes(trimmed)) return false;
-    collection.push(trimmed);
-    return true;
-  };
-
-  const registerTiktokPixelHint = (value: any, detected: string[]) => {
-    if (addValue(hints, value)) {
-      addValue(hints, 'tiktok');
-      detected.push(String(value).trim());
-      return true;
-    }
-    return false;
-  };
-
-  addValue(
-    sources,
-    pickByKeys(
-      body,
-      [
-        'utm_source',
-        'utmSource',
-        'trackingparameters.utm_source',
-        'traffic_source',
-        'campaign_source',
-        'source_platform',
-        'marketing_source',
-      ],
-      (v) => typeof v === 'string' && v.trim().length > 0,
-    ) as string | null,
-  );
-
-  addValue(
-    hints,
-    pickByKeys(
-      body,
-      [
-        'traffic_channel',
-        'trafficChannel',
-        'utm_channel',
-        'utmChannel',
-        'trackingparameters.utm_channel',
-        'channel',
-      ],
-      (v) => typeof v === 'string' && v.trim().length > 0,
-    ) as string | null,
-  );
-
-  addValue(
-    mediums,
-    pickByKeys(
-      body,
-      [
-        'utm_medium',
-        'utmMedium',
-        'trackingparameters.utm_medium',
-        'traffic_medium',
-        'campaign_medium',
-        'marketing_medium',
-      ],
-      (v) => typeof v === 'string' && v.trim().length > 0,
-    ) as string | null,
-  );
-
-  addValue(
-    hints,
-    pickByKeys(
-      body,
-      [
-        'utm_campaign',
-        'utmCampaign',
-        'trackingparameters.utm_campaign',
-        'campaign',
-        'campaign_name',
-        'campaignName',
-      ],
-      (v) => typeof v === 'string' && v.trim().length > 0,
-    ) as string | null,
-  );
-
-  addValue(
-    hints,
-    pickByKeys(body, ['adset', 'adset_name', 'adsetName', 'adgroup', 'adgroup_name', 'adgroupName', 'ad_name', 'adName'], (v) =>
-      typeof v === 'string' && v.trim().length > 0
-    ) as string | null
-  );
-
-  const detectedTikTokPixels: string[] = [];
-
-  for (const key of ['tiktok_pixel_id', 'tt_pixel_id']) {
-    registerTiktokPixelHint(
-      pickByKeys(body, [key], (v) => typeof v === 'string' && v.trim().length > 0) as string | null,
-      detectedTikTokPixels
-    );
-  }
-
-  const genericPixelId = pickByKeys(
-    body,
-    ['pixel_id', 'pixelId'],
-    (v) => typeof v === 'string' && v.trim().length > 0
-  ) as string | null;
-
-  if (genericPixelId && /^(tt|bt)/i.test(genericPixelId.trim())) {
-    registerTiktokPixelHint(genericPixelId, detectedTikTokPixels);
-  }
-
-  addValue(
-    hints,
-    pickByKeys(body, ['pixel', 'pixel_id', 'pixelId'], (v) =>
-      typeof v === 'string' && v.trim().length > 0
-    ) as string | null
-  );
-
-  if (detectedTikTokPixels.length) {
-    console.log('[kiwify-webhook] detected TikTok pixel hints', {
-      tiktokPixelIds: detectedTikTokPixels,
-    });
-  }
-
-  for (const key of ['ttclid', 'fbclid', 'gclid', 'msclkid', 'kwai_adid', 'kwaiAdId']) {
-    addValue(
-      hints,
-      pickByKeys(body, [key], (v) => typeof v === 'string' && v.trim().length > 0) as string | null
-    );
-  }
-
-  if (params) {
-    const param = (key: string) => params.get(key);
-    addValue(sources, param('utm_source'));
-    addValue(sources, param('source'));
-    addValue(sources, param('utm_platform'));
-    addValue(sources, param('utm_channel'));
-    addValue(sources, param('traffic_source'));
-    addValue(mediums, param('utm_medium'));
-    addValue(mediums, param('medium'));
-    addValue(mediums, param('traffic_medium'));
-    addValue(hints, param('utm_campaign'));
-    addValue(hints, param('utm_content'));
-    addValue(hints, param('utm_term'));
-    addValue(hints, param('utm_id'));
-    addValue(hints, param('utm_source_platform'));
-    addValue(hints, param('adset_name'));
-    addValue(hints, param('campaign_name'));
-    addValue(hints, param('ttclid'));
-    addValue(hints, param('fbclid'));
-    addValue(hints, param('gclid'));
-    addValue(hints, param('msclkid'));
-    addValue(hints, param('kwai_adid'));
-  }
-
-  const context = buildTrafficContext([...sources, ...mediums, ...hints], params);
-  const mediumContext = buildTrafficContext([...mediums, ...hints], params);
-
-  let channel: string | null = null;
-  for (const candidate of [...sources, ...hints]) {
-    channel = resolveChannelFromValue(candidate);
-    if (channel) break;
-  }
-
-  if (!channel) {
-    channel = detectChannelFromContext(context);
-
-    if (hasTrafficHint(context, ['tiktok', 'ttclid', 'ttad', 'ttadgroup', 'tiktokads'])) channel = 'tiktok';
-    else if (hasTrafficHint(context, ['instagram'])) channel = 'instagram';
-    else if (hasTrafficHint(context, ['facebook', 'meta', 'fbclid', 'fbads'])) channel = 'facebook';
-    else if (hasTrafficHint(context, ['google', 'gclid', 'adwords', 'googleads'])) channel = 'google';
-    else if (hasTrafficHint(context, ['bing', 'msclkid', 'microsoft'])) channel = 'bing';
-    else if (hasTrafficHint(context, ['taboola'])) channel = 'taboola';
-    else if (hasTrafficHint(context, ['kwai'])) channel = 'kwai';
-    else if (hasTrafficHint(context, ['pinterest', 'pin'])) channel = 'pinterest';
-    else if (hasTrafficHint(context, ['snapchat', 'snap'])) channel = 'snapchat';
-    else if (hasTrafficHint(context, ['twitter', 'xads', 'x com'])) channel = 'twitter';
-    else if (hasTrafficHint(context, ['linkedin'])) channel = 'linkedin';
-    else if (hasTrafficHint(context, ['youtube'])) channel = 'youtube';
-    else if (hasTrafficHint(context, ['email', 'newsletter'])) channel = 'email';
-    else if (hasTrafficHint(context, ['whatsapp', 'wa'])) channel = 'whatsapp';
-  }
-
-  const classification = detectTrafficClassification(mediumContext, params);
-
-  const candidateParts: string[] = [];
-  if (channel) {
-    candidateParts.push(channel);
-  }
-  if (
-    classification &&
-    classification !== 'unknown' &&
-    !candidateParts.includes(classification)
-  ) {
-    candidateParts.push(classification);
-  }
-
-  let candidate = joinTrafficParts(candidateParts);
-
-  if (!candidate) {
-    for (const value of sources) {
-      const normalized = resolveChannelFromValue(value);
-      if (normalized) {
-        candidate = normalized;
-        break;
-      }
-    }
-  }
-
-  if (!candidate && classification && classification !== 'unknown') {
-    candidate = joinTrafficParts([classification]);
-  }
-
-  let resolvedCandidate: string | null;
-
-  if (!candidate) {
-    resolvedCandidate = existingTrafficSource ?? null;
-  } else if (existingTrafficSource && existingTrafficSource !== 'unknown') {
-    const existingSpecific = existingTrafficSource.includes('.');
-    const candidateSpecific = candidate.includes('.');
-    resolvedCandidate = !candidateSpecific && existingSpecific ? existingTrafficSource : candidate;
-  } else {
-    resolvedCandidate = candidate;
-  }
-
-  if (manualReminder === 'email') {
-    const normalizeCandidate = (value: string | null | undefined): string | null => {
-      if (typeof value !== 'string') return null;
-      const normalized = value.trim().toLowerCase();
-      if (!normalized || normalized === 'unknown') return null;
-      return normalized;
-    };
-
-    const resolvedNormalized = normalizeCandidate(resolvedCandidate);
-    const existingNormalized = normalizeCandidate(existingTrafficSource);
-    const candidateNormalized = normalizeCandidate(candidate);
-
-    const isEmailLike = (value: string | null) =>
-      !value ? false : value === 'email' || value.endsWith('.email');
-
-    const baseCandidate =
-      (resolvedNormalized && !isEmailLike(resolvedNormalized) && resolvedNormalized) ||
-      (existingNormalized && !isEmailLike(existingNormalized) && existingNormalized) ||
-      (candidateNormalized && !isEmailLike(candidateNormalized) && candidateNormalized) ||
-      (existingNormalized && existingNormalized !== 'email' && existingNormalized) ||
-      null;
-
-    const baseParts = baseCandidate
-      ? baseCandidate
-          .split('.')
-          .map((part) => part.trim())
-          .filter((part) => part && part !== 'email')
-      : [];
-
-    if (!baseParts.length) {
-      baseParts.push('organic');
-    }
-
-    const manualSource = joinTrafficParts([...baseParts, 'email']);
-    return manualSource ?? 'organic.email';
-  }
-
-  return resolvedCandidate;
-}
 
 function findDiscountDeep(obj: any): string | null {
   const byKey = pickByKeys(
